Update and rename createtablecommand.txt to MySQLcommands.txt
These are the commands used in the MySQL database. The first command creates a table in the database that contains the username and password of a user. It also uses the username as the unique key that identifies a record in the database. The second command sets the default username to "John." This was done so that it was possible to demo the "POST" method and the current functionality it has, which is that it can create a new record in the database with a password of our choosing, but it currently can not create an account that has a username of our choosing.