Update MySQLcommands.txt
These are the commands used in the MySQL command line. The first command creates a table in the database that contains the username and password of a user. It also uses the username as the unique key that identifies a record in the table. The second command sets the default username to "John." This was done so that it was possible to demo the "POST" method and the current functionality it has, which is that it can create a new record in the database with a password of our choosing, but it currently can not create an account that has a username of our choosing and can only have the username "John." The third command creates two records directly through the MySQL command line rather than create them with the POST method so we can properly test the other methods despite the previously mentioned issue.